refactor(platform): extract startup entrypoint
